--- arkworks/src/arkmsm/lib.rs
@@ -1,8 +1,0 @@
-pub mod bitmap;
-pub mod glv;
-pub mod msm;
-pub mod types;
-pub mod utils;
-
-mod batch_adder;
-mod bucket_msm;
