+++ arkworks/src/fft_g1.rs
@@ -6,7 +6,10 @@ use crate::kzg_types::{ArkFr, ArkG1, ArkG1Affine};
 use crate::kzg_types::ArkG1ProjAddAffine;
 
 #[cfg(feature = "parallel")]
-use kzg::msm::tilling_parallel_pippinger::tiling_parallel_pippinger;
+use kzg::msm::tiling_parallel_pippenger::tiling_parallel_pippenger;
+
+#[cfg(not(feature = "parallel"))]
+use kzg::msm::arkmsm::arkmsm_msm::VariableBaseMSM;
 
 use ark_ff::BigInteger256;
 
@@ -38,7 +41,7 @@ pub fn g1_linear_combination(out: &mut ArkG1, points: &[ArkG1], scalars: &[ArkFr
                 .collect::<Vec<_>>()
         };
 
-        *out = tiling_parallel_pippinger(&ark_points, ark_scalars.as_slice());
+        *out = tiling_parallel_pippenger(&ark_points, ark_scalars.as_slice());
     }
 
     #[cfg(not(feature = "parallel"))]
@@ -50,7 +53,7 @@ pub fn g1_linear_combination(out: &mut ArkG1, points: &[ArkG1], scalars: &[ArkFr
                 .map(|scalar| Scalar256::from_u64(BigInteger256::from(scalar.fr).0))
                 .collect::<Vec<_>>()
         };
-        *out = kzg::msm::arkmsm_msm::VariableBaseMSM::multi_scalar_mul::<_, _, _, ArkG1ProjAddAffine>(
+        *out = VariableBaseMSM::multi_scalar_mul::<_, _, _, ArkG1ProjAddAffine>(
             &ark_points,
             &ark_scalars,
         );
