+++ blst/src/types/g1.rs
@@ -5,6 +5,7 @@ use core::ptr;
 use alloc::format;
 use alloc::string::String;
 use alloc::string::ToString;
+use alloc::vec::Vec;
 
 use blst::p1_affines;
 use blst::{
