+++ kzg/src/msm/arkmsm/arkmsm_msm.rs
@@ -1,9 +1,12 @@
 use crate::{
-    common_utils::log2_u64, msm::bucket_msm::BucketMSM, msm::glv::decompose,
-    msm::types::G1_SCALAR_SIZE_GLV, Fr, G1Affine, G1Fp, G1ProjAddAffine, Scalar256, G1,
+    common_utils::log2_u64,
+    msm::arkmsm::bucket_msm::BucketMSM,
+    msm::arkmsm::glv::decompose,
+    msm::types::{G1_SCALAR_SIZE, G1_SCALAR_SIZE_GLV},
+    Fr, G1Affine, G1Fp, G1ProjAddAffine, Scalar256, G1,
 };
 
-use super::types::G1_SCALAR_SIZE;
+use alloc::vec;
 
 pub struct VariableBaseMSM;
