+++ kzg/src/msm/arkmsm/bucket_msm.rs
@@ -1,55 +1,62 @@
-use ark_bls12_381::g1::Config as G1Parameters;
-use ark_bls12_381::G1Affine;
-use ark_ec::AffineRepr;
-use ark_ec::{
-    models::short_weierstrass::SWCurveConfig as Parameters,
-    short_weierstrass::{Affine, Projective},
-    Group,
-};
-use ark_std::Zero;
-use std::{any::TypeId, ops::AddAssign};
+use core::marker::PhantomData;
 
 #[cfg(feature = "parallel")]
 use rayon::prelude::*;
 
+use alloc::vec;
+use alloc::vec::Vec;
+
 use crate::{
-    arkmsm::batch_adder::BatchAdder,
-    arkmsm::bitmap::Bitmap,
-    arkmsm::glv::endomorphism,
-    arkmsm::types::{GROUP_SIZE, GROUP_SIZE_IN_BITS},
+    cfg_into_iter,
+    msm::arkmsm::{batch_adder::BatchAdder, bitmap::Bitmap, glv::endomorphism},
+    msm::types::{GROUP_SIZE, GROUP_SIZE_IN_BITS},
+    G1Affine, G1Fp, G1ProjAddAffine, G1,
 };
 
-pub struct BucketMSM<P: Parameters> {
-    num_windows: u32,
-    window_bits: u32,
-    bucket_bits: u32,
-    max_batch_cnt: u32, // max slices allowed in a batch
-    max_collision_cnt: u32,
-    buckets: Vec<Affine<P>>, // size (num_windows << window_bits) * 2
+pub struct BucketMSM<
+    TG1: G1,
+    TG1Fp: G1Fp,
+    TG1Affine: G1Affine<TG1, TG1Fp>,
+    TProjAddAffine: G1ProjAddAffine<TG1, TG1Fp, TG1Affine>,
+> {
+    pub num_windows: u32,
+    pub window_bits: u32,
+    pub bucket_bits: u32,
+    pub max_batch_cnt: u32, // max slices allowed in a batch
+    pub max_collision_cnt: u32,
+    pub buckets: Vec<TG1Affine>, // size (num_windows << window_bits) * 2
 
     // current batch state
-    bitmap: Bitmap,
-    batch_buckets_and_points: Vec<(u32, u32)>,
-    collision_buckets_and_points: Vec<(u32, Affine<P>)>,
-    cur_points: Vec<Affine<P>>, // points of current batch, size batch_size
+    pub bitmap: Bitmap,
+    pub batch_buckets_and_points: Vec<(u32, u32)>,
+    pub collision_buckets_and_points: Vec<(u32, TG1Affine)>,
+    pub cur_points: Vec<TG1Affine>, // points of current batch, size batch_size
 
     // batch affine adder
-    batch_adder: BatchAdder<P>,
+    pub batch_adder: BatchAdder<TG1, TG1Fp, TG1Affine>,
+    _p: PhantomData<TProjAddAffine>,
 }
 
-impl<P: Parameters> BucketMSM<P> {
+impl<
+        TG1: G1,
+        TG1Fp: G1Fp,
+        TG1Affine: G1Affine<TG1, TG1Fp>,
+        TProjAddAffine: G1ProjAddAffine<TG1, TG1Fp, TG1Affine>,
+    > BucketMSM<TG1, TG1Fp, TG1Affine, TProjAddAffine>
+{
     pub fn new(
         scalar_bits: u32,
         window_bits: u32,
         max_batch_cnt: u32,     // default: 4096
         max_collision_cnt: u32, // default: 128
-    ) -> BucketMSM<P> {
+    ) -> Self {
+        // TODO: Check if these can be turned into consts
         let num_windows = (scalar_bits + window_bits - 1) / window_bits;
-        let batch_size = std::cmp::max(8192, max_batch_cnt);
+        let batch_size = core::cmp::max(8192, max_batch_cnt);
         let bucket_bits = window_bits - 1; // half buckets needed because of signed-bucket-index
         let bucket_size = num_windows << bucket_bits;
         // size of batch_adder will be the max of batch_size and num_windows * groups per window
-        let batch_adder_size = std::cmp::max(batch_size, bucket_size >> GROUP_SIZE_IN_BITS);
+        let batch_adder_size = core::cmp::max(batch_size, bucket_size >> GROUP_SIZE_IN_BITS);
 
         BucketMSM {
             num_windows,
@@ -57,30 +64,26 @@ impl<P: Parameters> BucketMSM<P> {
             bucket_bits,
             max_batch_cnt,
             max_collision_cnt,
-            buckets: vec![Affine::<P>::zero(); bucket_size as usize],
+            buckets: vec![TG1Affine::ZERO; bucket_size as usize],
 
             bitmap: Bitmap::new(bucket_size as usize / 32),
             batch_buckets_and_points: Vec::with_capacity(batch_size as usize),
             collision_buckets_and_points: Vec::with_capacity(max_collision_cnt as usize),
-            cur_points: vec![Affine::<P>::zero(); batch_size as usize],
+            cur_points: vec![TG1Affine::ZERO; batch_size as usize],
 
             batch_adder: BatchAdder::new(batch_adder_size as usize),
+            _p: PhantomData,
         }
     }
 
     pub fn process_point_and_slices_glv(
         &mut self,
-        point: &Affine<P>,
+        point: &TG1Affine,
         normal_slices: &[u32],
         phi_slices: &[u32],
         is_neg_scalar: bool,
         is_neg_normal: bool,
     ) {
-        assert_eq!(
-            TypeId::of::<P>(),
-            TypeId::of::<G1Parameters>(),
-            "glv is only supported for ark_bls12_381"
-        );
         assert!(
             self.num_windows as usize == normal_slices.len()
                 && normal_slices.len() == phi_slices.len(),
@@ -93,10 +96,11 @@ impl<P: Parameters> BucketMSM<P> {
         let mut p = *point; // copy
 
         if is_neg_scalar {
-            p.y = -p.y
+            p.y_mut().neg_assign();
         };
+        // TODO: Can be replaced with XOR?
         if is_neg_normal {
-            p.y = -p.y
+            p.y_mut().neg_assign();
         };
 
         self.cur_points.push(p);
@@ -107,7 +111,7 @@ impl<P: Parameters> BucketMSM<P> {
             }
         }
 
-        p.y = -p.y;
+        p.y_mut().neg_assign();
 
         self.cur_points.push(p);
         for (win, normal_slice) in normal_slices.iter().enumerate() {
@@ -121,13 +125,13 @@ impl<P: Parameters> BucketMSM<P> {
         }
 
         // process phi slices
-        p.y = -p.y;
+        p.y_mut().neg_assign();
         if is_neg_normal {
-            p.y = -p.y;
+            p.y_mut().neg_assign();
         }
 
         // this isn't the cleanest of doing this, we'd better figure out a way to do this at compile time
-        let p_g1: &mut G1Affine = unsafe { &mut *(std::ptr::addr_of_mut!(p) as *mut G1Affine) };
+        let p_g1: &mut TG1Affine = &mut p;
         endomorphism(p_g1);
 
         self.cur_points.push(p);
@@ -138,7 +142,7 @@ impl<P: Parameters> BucketMSM<P> {
             }
         }
 
-        p.y = -p.y;
+        p.y_mut().neg_assign();
 
         self.cur_points.push(p);
         for (win, phi_slice) in phi_slices.iter().enumerate() {
@@ -152,7 +156,7 @@ impl<P: Parameters> BucketMSM<P> {
         }
     }
 
-    pub fn process_point_and_slices(&mut self, point: &Affine<P>, slices: &[u32]) {
+    pub fn process_point_and_slices(&mut self, point: &TG1Affine, slices: &[u32]) {
         assert!(
             self.num_windows as usize == slices.len(),
             "slices.len() {} should equal num_windows {}",
@@ -169,7 +173,7 @@ impl<P: Parameters> BucketMSM<P> {
         }
 
         let mut neg_p = *point;
-        neg_p.y = -neg_p.y;
+        neg_p.y_mut().neg_assign();
 
         self.cur_points.push(neg_p);
         for (win, slice) in slices.iter().enumerate() {
@@ -249,12 +253,12 @@ impl<P: Parameters> BucketMSM<P> {
         self.cur_points.push(slicing_point.unwrap());
     }
 
-    pub fn batch_reduce(&mut self) -> Projective<P> {
+    pub fn batch_reduce(&mut self) -> TG1 {
         let window_starts: Vec<_> = (0..self.num_windows as usize).collect();
         let num_groups =
             (self.num_windows as usize) << (self.bucket_bits as usize - GROUP_SIZE_IN_BITS);
-        let mut running_sums: Vec<_> = vec![Affine::<P>::zero(); num_groups];
-        let mut sum_of_sums: Vec<_> = vec![Affine::<P>::zero(); num_groups];
+        let mut running_sums: Vec<_> = vec![TG1Affine::ZERO; num_groups];
+        let mut sum_of_sums: Vec<_> = vec![TG1Affine::ZERO; num_groups];
 
         // calculate running sum and sum of sum for each group
         for i in (0..GROUP_SIZE).rev() {
@@ -270,7 +274,7 @@ impl<P: Parameters> BucketMSM<P> {
             self.batch_adder.batch_add(&mut sum_of_sums, &running_sums);
         }
 
-        let sum_by_window: Vec<Projective<P>> = ark_std::cfg_into_iter!(window_starts)
+        let sum_by_window: Vec<TG1> = cfg_into_iter!(window_starts)
             .map(|w_start| {
                 let group_start = w_start << (self.bucket_bits as usize - GROUP_SIZE_IN_BITS);
                 let group_end = (w_start + 1) << (self.bucket_bits as usize - GROUP_SIZE_IN_BITS);
@@ -284,125 +288,47 @@ impl<P: Parameters> BucketMSM<P> {
         self.intra_window_reduce(&sum_by_window)
     }
 
-    fn inner_window_reduce(
-        &self,
-        running_sums: &[Affine<P>],
-        sum_of_sums: &[Affine<P>],
-    ) -> Projective<P> {
-        self.calc_sum_of_sum_total(sum_of_sums) + self.calc_running_sum_total(running_sums)
+    fn inner_window_reduce(&self, running_sums: &[TG1Affine], sum_of_sums: &[TG1Affine]) -> TG1 {
+        self.calc_sum_of_sum_total(sum_of_sums)
+            .add_or_dbl(&self.calc_running_sum_total(running_sums))
     }
 
-    fn calc_running_sum_total(&self, running_sums: &[Affine<P>]) -> Projective<P> {
-        let mut running_sum_total = Projective::<P>::zero();
+    fn calc_running_sum_total(&self, running_sums: &[TG1Affine]) -> TG1 {
+        let mut running_sum_total = TG1::ZERO;
         for (i, running_sum) in running_sums.iter().enumerate().skip(1) {
             for _ in 0..i {
-                running_sum_total.add_assign(running_sum);
+                TProjAddAffine::add_or_double_assign_affine(&mut running_sum_total, running_sum);
             }
         }
 
         for _ in 0..GROUP_SIZE_IN_BITS {
-            running_sum_total.double_in_place();
+            running_sum_total.dbl_assign();
         }
         running_sum_total
     }
 
-    fn calc_sum_of_sum_total(&self, sum_of_sums: &[Affine<P>]) -> Projective<P> {
-        let mut sum = Projective::<P>::zero();
-        sum_of_sums.iter().for_each(|p| sum.add_assign(p));
+    fn calc_sum_of_sum_total(&self, sum_of_sums: &[TG1Affine]) -> TG1 {
+        let mut sum = TG1::ZERO;
+        sum_of_sums
+            .iter()
+            .for_each(|p| TProjAddAffine::add_or_double_assign_affine(&mut sum, p));
         sum
     }
 
-    fn intra_window_reduce(&self, window_sums: &[Projective<P>]) -> Projective<P> {
-        // We store the sum for the lowest window.
-        let lowest = *window_sums.first().unwrap();
-
-        // We're traversing windows from high to low.
-        lowest
-            + window_sums
+    fn intra_window_reduce(&mut self, window_sums: &[TG1]) -> TG1 {
+        // Traverse windows from high to low
+        let lowest = window_sums.first().unwrap();
+        lowest.add(
+            &window_sums[1..]
                 .iter()
-                .skip(1)
                 .rev()
-                .fold(Projective::<P>::zero(), |mut total, sum_i| {
-                    total += sum_i;
+                .fold(TG1::ZERO, |mut total, sum_i| {
+                    total.add_assign(sum_i);
                     for _ in 0..self.window_bits {
-                        total.double_in_place();
+                        total.dbl_assign();
                     }
                     total
-                })
-    }
-}
-
-#[cfg(test)]
-mod bucket_msm_tests {
-    use super::*;
-    use ark_bls12_381::{G1Affine, G1Projective};
-    use ark_std::UniformRand;
-
-    #[test]
-    fn test_process_point_and_slices_deal_two_points() {
-        let window_bits = 15u32;
-        let mut bucket_msm = BucketMSM::new(30u32, window_bits, 128u32, 4096u32);
-        let mut rng = ark_std::test_rng();
-        let p_prj = G1Projective::rand(&mut rng);
-        let q_prj = G1Projective::rand(&mut rng);
-        let p = G1Affine::from(p_prj);
-        let q = G1Affine::from(q_prj);
-
-        bucket_msm.process_point_and_slices(&p, &[1u32, 3u32]);
-        bucket_msm.process_point_and_slices(&q, &[2u32, 3u32]);
-        bucket_msm.process_complete();
-        assert_eq!(bucket_msm.buckets[0], p);
-        assert_eq!(bucket_msm.buckets[1], q);
-        assert_eq!(bucket_msm.buckets[2 + (1 << bucket_msm.bucket_bits)], p + q);
-    }
-
-    #[test]
-    fn test_process_point_and_slices_deal_three_points() {
-        let window_bits = 15u32;
-        let mut bucket_msm = BucketMSM::new(45u32, window_bits, 128u32, 4096u32);
-        let mut rng = ark_std::test_rng();
-        let p_prj = G1Projective::rand(&mut rng);
-        let q_prj = G1Projective::rand(&mut rng);
-        let r_prj = G1Projective::rand(&mut rng);
-        let p = G1Affine::from(p_prj);
-        let q = G1Affine::from(q_prj);
-        let r = G1Affine::from(r_prj);
-
-        bucket_msm.process_point_and_slices(&p, &[1u32, 3u32, 4u32]);
-        bucket_msm.process_point_and_slices(&q, &[2u32, 3u32, 4u32]);
-        bucket_msm.process_point_and_slices(&r, &[2u32, 3u32, 5u32]);
-        bucket_msm.process_complete();
-        assert_eq!(bucket_msm.buckets[0], p);
-        assert_eq!(bucket_msm.buckets[1], q + r);
-        assert_eq!(
-            bucket_msm.buckets[2 + (1 << bucket_msm.bucket_bits)],
-            p + q + r
-        );
-        assert_eq!(bucket_msm.buckets[3 + (2 << bucket_msm.bucket_bits)], p + q);
-        assert_eq!(bucket_msm.buckets[4 + (2 << bucket_msm.bucket_bits)], r);
-    }
-
-    #[test]
-    fn test_process_point_and_slices_glv_deal_two_points() {
-        let window_bits = 15u32;
-        let mut bucket_msm = BucketMSM::new(30u32, window_bits, 128u32, 4096u32);
-        let mut rng = ark_std::test_rng();
-        let p_prj = G1Projective::rand(&mut rng);
-        let q_prj = G1Projective::rand(&mut rng);
-        let mut p = G1Affine::from(p_prj);
-        let mut q = G1Affine::from(q_prj);
-
-        bucket_msm.process_point_and_slices_glv(&p, &[1u32, 3u32], &[4u32, 6u32], false, false);
-        bucket_msm.process_point_and_slices_glv(&q, &[2u32, 3u32], &[5u32, 6u32], false, false);
-        bucket_msm.process_complete();
-        assert_eq!(bucket_msm.buckets[0], p);
-        assert_eq!(bucket_msm.buckets[1], q);
-        assert_eq!(bucket_msm.buckets[2 + (1 << bucket_msm.bucket_bits)], p + q);
-
-        endomorphism(&mut p);
-        endomorphism(&mut q);
-        assert_eq!(bucket_msm.buckets[3], p);
-        assert_eq!(bucket_msm.buckets[4], q);
-        assert_eq!(bucket_msm.buckets[5 + (1 << bucket_msm.bucket_bits)], p + q);
+                }),
+        )
     }
 }
