+++ kzg/src/msm/arkmsm/mod.rs
@@ -1,7 +1,5 @@
+pub mod arkmsm_msm;
 pub mod batch_adder;
 pub mod bitmap;
 pub mod bucket_msm;
 pub mod glv;
-pub mod msm;
-pub mod types;
-pub mod utils;
