+++ kzg/src/msm/mod.rs
@@ -1,11 +1,10 @@
-pub mod arkmsm_msm;
-pub mod batch_adder;
-pub mod bitmap;
-pub mod bucket_msm;
+pub mod arkmsm;
 pub mod cell;
-pub mod glv;
-pub mod parallel_pippinger;
-pub mod thread_pool;
-pub mod tilling_parallel_pippinger;
-pub mod tilling_pippinger_ops;
+pub mod msm;
+pub mod tiling_pippenger_ops;
 pub mod types;
+
+#[cfg(feature = "parallel")]
+pub mod thread_pool;
+#[cfg(feature = "parallel")]
+pub mod tiling_parallel_pippenger;
